Extract common functions into StatusReporterTest
