+++ java/clients/src/test/java/sleeper/status/report/compactionjobstatus/StatusReporterTest.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.status.report.compactionjobstatus;
+
+import org.apache.commons.io.IOUtils;
+import org.junit.Before;
+import sleeper.compaction.job.CompactionJobTestDataHelper;
+import sleeper.status.report.compactionjob.CompactionJobStatusReporter;
+import sleeper.status.report.filestatus.FilesStatusReportTest;
+
+import java.io.IOException;
+import java.net.URL;
+import java.nio.charset.Charset;
+import java.util.Objects;
+
+public abstract class StatusReporterTest {
+    protected CompactionJobTestDataHelper dataHelper;
+    protected CompactionJobStatusReporter statusReporter;
+    @Before
+    public void setup() {
+        statusReporter = new CompactionJobStatusReporter();
+        dataHelper = new CompactionJobTestDataHelper();
+    }
+    protected static String example(String path) throws IOException {
+        URL url = FilesStatusReportTest.class.getClassLoader().getResource(path);
+        return IOUtils.toString(Objects.requireNonNull(url), Charset.defaultCharset());
+    }
+}
